refactor: rename specifications folder to documents
