+++ packages/galaxy/src/index.ts
@@ -1,4 +1,4 @@
-import galaxySpecBase from './specifications/3.1.yaml'
+import galaxySpecBase from './documents/3.1.yaml'
 
 const galaxySpec: Record<string, any> = galaxySpecBase
