+++ packages/galaxy/tests/validate.test.ts
@@ -1,7 +1,7 @@
 import { validate } from '@scalar/openapi-parser'
 import { describe, expect, it } from 'vitest'
 
-import galaxy from '../src/specifications/3.1.yaml?raw'
+import galaxy from '../src/documents/3.1.yaml?raw'
 
 describe('validate', () => {
   it('is valid', async () => {
