+++ packages/galaxy/tests/yaml.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import galaxy from '../src/specifications/3.1.yaml?raw'
+import galaxy from '../src/documents/3.1.yaml?raw'
 
 describe('yaml', () => {
   it('has OpenAPI version', () => {
